requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
Django==4.2.16
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.116
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.44.0
orjson==3.10.7
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.10.0
zipp==3.20.1
